added cast to int in SongDao

+++ src/main/java/com/musicDB/dao/SongDAOImpl.java
@@ -46,7 +46,7 @@ public class SongDAOImpl implements SongDAO {
         Session currentSession = sessionFactory.getCurrentSession();
 
         // Set song id to 0 so that DAO inserts a new customer instead of updating an existing one
-        song.setId(0);
+        song.setId((long) 0);
 
         long songId = (long) currentSession.save(song);
         song.setId(songId);
